Truncate to int to avoid subpixel rendering.
On browsers that don't accelerate canvas, subpixel rendering is done in
software and is a lot slower. Rendering on whole pixels can allow software
renderers to just do a simple pixel copy.

On my system (Linux, Firefox 4 beta 10) this change significantly speeds
up the canvas demo and improves my overall score from 81 to 143 sprites.
This seems to be supported by the results of this benchmark:
http://jsperf.com/drawimage-whole-pixels